refactor: renames to tracker import strategy
