+++ dhis-2/dhis-web/dhis-web-api/src/main/java/org/hisp/dhis/webapi/strategy/tracker/imports/TrackerImportStrategy.java
@@ -25,32 +25,15 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.webapi.strategy.tracker.imports.impl;
+package org.hisp.dhis.webapi.strategy.tracker.imports;
 
-import lombok.RequiredArgsConstructor;
-
-import org.hisp.dhis.tracker.TrackerImportService;
 import org.hisp.dhis.tracker.report.TrackerImportReport;
 import org.hisp.dhis.webapi.controller.tracker.TrackerImportReportRequest;
-import org.hisp.dhis.webapi.strategy.tracker.imports.ImportStrategy;
-import org.springframework.stereotype.Component;
 
 /**
  * @author Luca Cambi <luca@dhis2.org>
  */
-@Component
-@RequiredArgsConstructor
-public class ImportAsyncFalseStrategyImpl implements ImportStrategy
+public interface TrackerImportStrategy
 {
-    private final TrackerImportService trackerImportService;
-
-    @Override
-    public TrackerImportReport importReport( TrackerImportReportRequest trackerImportReportRequest )
-    {
-        TrackerImportReport trackerImportReport = trackerImportService
-            .importTracker( trackerImportReportRequest.getTrackerImportParams() );
-
-        return trackerImportService.buildImportReport( trackerImportReport,
-            trackerImportReportRequest.getTrackerBundleReportMode() );
-    }
+    TrackerImportReport importReport( TrackerImportReportRequest trackerImportReportRequest );
 }
